Back-End e Banco de dados funcionando no web

+++ Back-End/src/main/java/com/tastycuisine/TastyCuisineV2/model/entity/Comentario.java
@@ -12,7 +12,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -42,7 +41,7 @@ public class Comentario {
     private Receita receita; 
     
     @Column(name = "Nota", nullable = false)
-    private Double nota;
+private Integer nota;
 
     @Builder.Default
     @Column(name = "Status_Comentarios",length =  20,nullable = false)
